+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Fixes
 
+* `AposInputBoolean` can now be `required` and have the value `false`.
+* Schema fields containing boolean filters can now list both `yes` and `no` choices according to available values in the database.
 * Fix attachment `getHeight()` and `getWidth()` template helpers by changing the assignment of the `attachment._crop` property.
 * Change assignment of `attachment._focalPoint` for consistency.